WIP - communication with backend
Issues:
* playback does not work--probably a naming issue
* legend & streamgraph colors are broken for playback of files
* streamgraph & sensor seem a bit too small on desktop browser
