--- README.md
@@ -1,9 +1,0 @@
-# Getting Started
-`simulate.js` contains a node server with pre-recorded UIC campus data.  Running
-`npm start` will launch both the node server and the app.  From there you can
-view it at `localhost:8000`.
-
-```
-npm install
-npm run dev
-```
