--- Server/1480454456047.txt
@@ -1,1 +1,0 @@
-{"name":"flare","lat":"41.8713644660119","lng":"-87.6484761742445","children":[{"name":"RF2449.188MHZ","size":"-91.68424"},{"name":"RF2401.938MHZ","size":"-98.22453"},{"name":"RF101.719MHZ","size":"-63.34301"},{"name":"RF105MHZ","size":"-68.30395"},{"name":"RF93.625MHZ","size":"-77.78832"}]}
