+++ index.html
@@ -20,12 +20,12 @@
 
      <svg id="legend"></svg>
 
-     <button class="mdl-button mdl-js-button mdl-button--icon" id="config">
-       <i class="material-icons">settings</i>
-     </button>
-
      <button class="mdl-button mdl-js-button mdl-button--icon" id="pause">
        <i class="material-icons">pause</i>
+     </button>
+
+     <button class="mdl-button mdl-js-button mdl-button--icon" id="record">
+       <i class="material-icons">fiber_manual_record</i>
      </button>
 
      <button class="mdl-button mdl-js-button mdl-button--icon" id="locked">
@@ -36,21 +36,10 @@
        <i class="material-icons">format_list_bulleted</i>
      </button>
 
-     <div class="demo-card-wide mdl-card mdl-shadow--2dp" id="settings">
-       <div class="mdl-card__title">
-         <h2 class="mdl-card__title-text">Settings</h2>
-       </div>
-       <div class="mdl-card__supporting-text">
-         <select id="data-source" >
-           <option value="sensor">Sensor</option>
-           <option value="campus3.json">Campus3</option>
-         </select>
-       </div>
-       <div class="mdl-card__actions mdl-card--border">
-         <a class="mdl-button mdl-button--colored mdl-js-button mdl-js-ripple-effect">
-           Save
-         </a>
-       </div>
+     <div id="source">
+       <select id="data-source">
+         <option value="live">live</option>
+       </select>
      </div>
 
      <div id="stream"></div>
