+++ start.sh
@@ -2,13 +2,13 @@ echo "Transpiling via babel"
 npm run build
 
 echo "Launching python server on ws://localhost:8088"
-python3 ./backend/fileCreator.py &
+cd Server
+python3 filecreator.py&
 FILECREATOR_PID=$!
-
-python3 ./backend/socketioServer.py &
+python3 socketioServer.py&
 SOCKETIOSERVER_PID=$!
-
+cd ../
 echo "Launching app on http://localhost:8000"
-live-server --port=8000
+python -m SimpleHTTPServer 8000
 kill $FILECREATOR_PID
 kill $SOCKETIOSERVER_PID
