+++ style.css
@@ -28,19 +28,6 @@ html, body, #map {
   background: rgba(255, 255, 255, 0) !important;
 }
 
-.demo-card-wide.mdl-card {
-  width: calc(100% - 20px);
-  padding: 0;
-}
-
-.demo-card-wide > .mdl-card__menu {
-  color: #fff;
-}
-
-#settings {
-  display: none;
-}
-
 #stream {
   height: 150px;
   opacity: 0.7;
@@ -51,12 +38,22 @@ html, body, #map {
   width: 100vw;
 }
 
-#config, #pause, #locked, #legend, #legend-icon {
+#record, #source, #pause, #locked, #legend, #legend-icon {
   background: white;
   border: 2px solid rgba(0,0,0,0.2);
   border-radius: 4px;
 }
 
-#config:hover, #pause:hover, #locked:hover, #legend-icon:hover {
+select {
+  width: 100%;
+  padding: 4%;
+  font-size: 14px;
+}
+
+option {
+  font-size: 14px;
+}
+
+#record:hover, #source:hover, #pause:hover, #locked:hover, #legend-icon:hover {
   background: #f4f4f4;
 }
